refactor to clean things up
