+++ projects/plots/plots/pages/page.py
@@ -2,6 +2,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from bokeh.layouts import LayoutDOM
+    from ledger.events import EventSet, RecoveredInjectionSet
     from plots.app import App
 
 
@@ -12,7 +13,9 @@ class Page:
     def initialize_sources(self) -> None:
         raise NotImplementedError
 
-    def update(self) -> None:
+    def update(
+        self, background: EventSet, foreground: RecoveredInjectionSet
+    ) -> None:
         return
 
     def get_layout(self) -> "LayoutDOM":
